Remove git integration
This package has a built-in server that allows you to work with
git-based templates, but we only use the actual conversion
functionality. Since the `nodegit` dependency is not maintained and
pulls in dependencies with active CVEs, the simplest thing to do is to
remove the git-based APIs from the service.